ChatBotProject/backend/requirements.txt
Flask
PyPDF2
google-generativeai
python-dotenv
requests